examples

+++ Day-06/conditionTrue.py
@@ -1,5 +1,4 @@
-# Prompt the user to input a number and convert it to a floating-point value.
-n = float(input("Input a number: "))
+n = int(input("Input a number: "))
 
 # Check if the value of n is equal to 1.
 # If the condition is True, print the message indicating the first day of the month.
